+++ Makefile
@@ -6,22 +6,22 @@ VERSION := $(shell git describe --tags 2> /dev/null || echo unknown)
 all: test
 
 cover:
-	go test -cover -tags="integration fullcoverage" $(TESTFLAGS)
+	go test ./... -cover -tags="integration fullcoverage" $(TESTFLAGS)
 
 fast-cover:
-	go test -cover -tags="integration" $(TESTFLAGS)
+	go test ./... -cover -tags="integration" $(TESTFLAGS)
 
 generate:
 	protoc --go_out=. gopilosa_pbuf/public.proto
 
 test:
-	go test $(TESTFLAGS)
+	go test ./... $(TESTFLAGS)
 
 test-all:
-	go test -tags=integration $(TESTFLAGS)
+	go test ./... -tags=integration $(TESTFLAGS)
 
 test-all-race:
-	go test -race -tags=integration $(TESTFLAGS)
+	go test ./... -race -tags=integration $(TESTFLAGS)
 
 release:
 	printf "package pilosa\nconst Version = \"$(VERSION)\"" > version.go
