--- imports.go
@@ -1,261 +1,0 @@
-// Copyright 2017 Pilosa Corp.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions
-// are met:
-//
-// 1. Redistributions of source code must retain the above copyright
-// notice, this list of conditions and the following disclaimer.
-//
-// 2. Redistributions in binary form must reproduce the above copyright
-// notice, this list of conditions and the following disclaimer in the
-// documentation and/or other materials provided with the distribution.
-//
-// 3. Neither the name of the copyright holder nor the names of its
-// contributors may be used to endorse or promote products derived
-// from this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
-// CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
-// INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
-// MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-// BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-// WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
-// DAMAGE.
-
-package pilosa
-
-import (
-	"bufio"
-	"fmt"
-	"io"
-	"strconv"
-	"strings"
-	"time"
-
-	"github.com/pkg/errors"
-)
-
-type Record interface {
-	Shard(shardWidth uint64) uint64
-	Less(other Record) bool
-}
-
-type recordSort []Record
-
-func (rc recordSort) Len() int {
-	return len(rc)
-}
-
-func (rc recordSort) Swap(i, j int) {
-	rc[i], rc[j] = rc[j], rc[i]
-}
-
-func (rc recordSort) Less(i, j int) bool {
-	return rc[i].Less(rc[j])
-}
-
-type RecordIterator interface {
-	NextRecord() (Record, error)
-}
-
-// Column defines a single Pilosa column.
-type Column struct {
-	RowID     uint64
-	ColumnID  uint64
-	RowKey    string
-	ColumnKey string
-	Timestamp int64
-}
-
-func (b Column) Shard(shardWidth uint64) uint64 {
-	return b.ColumnID / shardWidth
-}
-
-func (b Column) Less(other Record) bool {
-	if ob, ok := other.(Column); ok {
-		if b.RowID == ob.RowID {
-			return b.ColumnID < ob.ColumnID
-		}
-		return b.RowID < ob.RowID
-	}
-	return false
-}
-
-type CSVFormat uint
-
-const (
-	CSVRowIDColumnID CSVFormat = iota
-	CSVRowIDColumnKey
-	CSVRowKeyColumnID
-	CSVRowKeyColumnKey
-	CSVColumnID
-	CSVColumnKey
-)
-
-func ColumnCSVUnmarshaller(format CSVFormat) CSVRecordUnmarshaller {
-	return ColumnCSVUnmarshallerWithTimestamp(format, "")
-}
-
-func ColumnCSVUnmarshallerWithTimestamp(format CSVFormat, timestampFormat string) CSVRecordUnmarshaller {
-	return func(text string) (Record, error) {
-		var err error
-		column := Column{}
-		parts := strings.Split(text, ",")
-		if len(parts) < 2 {
-			return nil, errors.New("Invalid CSV line")
-		}
-
-		hasRowKey := format == CSVRowKeyColumnID || format == CSVRowKeyColumnKey
-		hasColumnKey := format == CSVRowIDColumnKey || format == CSVRowKeyColumnKey
-
-		if hasRowKey {
-			column.RowKey = parts[0]
-		} else {
-			column.RowID, err = strconv.ParseUint(parts[0], 10, 64)
-			if err != nil {
-				return nil, errors.New("Invalid row ID")
-			}
-		}
-
-		if hasColumnKey {
-			column.ColumnKey = parts[1]
-		} else {
-			column.ColumnID, err = strconv.ParseUint(parts[1], 10, 64)
-			if err != nil {
-				return nil, errors.New("Invalid column ID")
-			}
-		}
-
-		timestamp := 0
-		if len(parts) == 3 {
-			if timestampFormat == "" {
-				timestamp, err = strconv.Atoi(parts[2])
-				if err != nil {
-					return nil, err
-				}
-			} else {
-				t, err := time.Parse(timestampFormat, parts[2])
-				if err != nil {
-					return nil, err
-				}
-				timestamp = int(t.Unix())
-			}
-		}
-		column.Timestamp = int64(timestamp)
-
-		return column, nil
-	}
-}
-
-type CSVRecordUnmarshaller func(text string) (Record, error)
-
-// CSVIterator reads records from a Reader.
-// Each line should contain a single record in the following form:
-// field1,field2,...
-type CSVIterator struct {
-	reader       io.Reader
-	line         int
-	scanner      *bufio.Scanner
-	unmarshaller CSVRecordUnmarshaller
-}
-
-// NewCSVIterator creates a CSVIterator from a Reader.
-func NewCSVIterator(reader io.Reader, unmarshaller CSVRecordUnmarshaller) *CSVIterator {
-	return &CSVIterator{
-		reader:       reader,
-		line:         0,
-		scanner:      bufio.NewScanner(reader),
-		unmarshaller: unmarshaller,
-	}
-}
-
-func NewCSVColumnIterator(format CSVFormat, reader io.Reader) *CSVIterator {
-	return NewCSVIterator(reader, ColumnCSVUnmarshaller(format))
-}
-
-func NewCSVColumnIteratorWithTimestampFormat(format CSVFormat, reader io.Reader, timestampFormat string) *CSVIterator {
-	return NewCSVIterator(reader, ColumnCSVUnmarshallerWithTimestamp(format, timestampFormat))
-}
-
-func NewCSVValueIterator(format CSVFormat, reader io.Reader) *CSVIterator {
-	return NewCSVIterator(reader, FieldValueCSVUnmarshaller(format))
-}
-
-// NextRecord iterates on lines of a Reader.
-// Returns io.EOF on end of iteration.
-func (c *CSVIterator) NextRecord() (Record, error) {
-	if ok := c.scanner.Scan(); ok {
-		c.line++
-		text := strings.TrimSpace(c.scanner.Text())
-		if text != "" {
-			rc, err := c.unmarshaller(text)
-			if err != nil {
-				return nil, fmt.Errorf("%s at line: %d", err.Error(), c.line)
-			}
-			return rc, nil
-		}
-	}
-	err := c.scanner.Err()
-	if err != nil {
-		return nil, err
-	}
-	return nil, io.EOF
-}
-
-// FieldValue represents the value for a column within a
-// range-encoded field.
-type FieldValue struct {
-	ColumnID  uint64
-	ColumnKey string
-	Value     int64
-}
-
-func (v FieldValue) Shard(shardWidth uint64) uint64 {
-	return v.ColumnID / shardWidth
-}
-
-func (v FieldValue) Less(other Record) bool {
-	if ov, ok := other.(FieldValue); ok {
-		return v.ColumnID < ov.ColumnID
-	}
-	return false
-}
-
-func FieldValueCSVUnmarshaller(format CSVFormat) CSVRecordUnmarshaller {
-	return func(text string) (Record, error) {
-		parts := strings.Split(text, ",")
-		if len(parts) < 2 {
-			return nil, errors.New("Invalid CSV")
-		}
-		value, err := strconv.ParseInt(parts[1], 10, 64)
-		if err != nil {
-			return nil, errors.New("Invalid value")
-		}
-		switch format {
-		case CSVColumnID:
-			columnID, err := strconv.ParseUint(parts[0], 10, 64)
-			if err != nil {
-				return nil, errors.New("Invalid column ID at line: %d")
-			}
-			return FieldValue{
-				ColumnID: uint64(columnID),
-				Value:    value,
-			}, nil
-		case CSVColumnKey:
-			return FieldValue{
-				ColumnKey: parts[0],
-				Value:     value,
-			}, nil
-		default:
-			return nil, fmt.Errorf("Invalid format: %d", format)
-		}
-	}
-}
